Extract PriorityQueue benchmarks into separate file
